move status & cleanup rpc structure
